requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
python-jose[cryptography]>=3.3.0

# Database & Storage
pymongo>=4.6.0
boto3>=1.34.0
redis>=5.0.1

# Image Processing (Python 3.12 compatible versions)
numpy>=1.26.0
Pillow>=10.1.0
opencv-python-headless>=4.8.1.78
imagehash>=4.3.1
scikit-learn>=1.3.2

# Job Queue
celery>=5.3.4
kombu>=5.3.4

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.2
aiofiles>=23.2.1

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0